refactor swaggerconfig, added fake information
